Fixing Blend Normal
The normal calculation has been fixed, and uses the contribution of each
component sphere to the resulting shape as the weight of the normal
contribution.

The surface becomes discontinuous where the two field functions overlap. This
effect is not fixed by removing the BVH, so it is caused by something else.