Root Finding Performance Boost
The root finder has been given a minor performance boost by only evaluating the
field function once per iteration. Furthermore, it checks if the field value
has evaluated to 0, and if it has, it terminates immediately. We are able to
turn the number of iterations down to 15 from 100, which also improves speed
without noticeable changes to the quality of the image.

The Blend function has been given some extra debug information.

We removed two empty lines from the sphere implementation for cleanliness.